encapsulate the methods inside a class
