--- src/handlers/displayEndTime.js
@@ -1,13 +1,0 @@
-
-
-import {pomo} from '../timer/data.js'
-
-
-
-export function displayEndTime(timestamp) {
-    const end = new Date(timestamp);
-    const hour = end.getHours();
-    const adjustedHour = hour > 12 ? hour - 12 : hour;
-    const minutes = end.getMinutes();
-    pomo.endTime.textContent = `Be Back At ${adjustedHour}:${minutes < 10 ? '0' : ''}${minutes}`;
-  }
