--- src/handlers/displayTimeLeft.js
@@ -1,13 +1,0 @@
-
-import {pomo} from '../timer/data.js'
-
-
-
-
-export function displayTimeLeft(seconds) {
-    const minutes = Math.floor(seconds / 60);
-    const remainderSeconds = seconds % 60;
-    const display = `${minutes}:${remainderSeconds < 10 ? '0' : '' }${remainderSeconds}`;
-    document.title = display;
-    pomo.timerDisplay.textContent = display;
-  }
